+++ app/Livewire/UploadedFileEdit.php
@@ -2,20 +2,76 @@
 
 namespace App\Livewire;
 
+use App\Facade\NotificationMessage;
+use App\Livewire\Forms\UploadedFileForm;
 use App\Models\UploadedFile;
-use Illuminate\Support\Facades\Storage;
+use App\NotificationMessage\Item;
+use App\NotificationMessage\ItemType;
+use Illuminate\Support\Facades\Log;
+use Illuminate\Support\Str;
+use Illuminate\Validation\ValidationException;
+use Livewire\Attributes\Validate;
 use Livewire\Component;
-use Symfony\Component\HttpFoundation\StreamedResponse;
+use Livewire\Features\SupportFileUploads\TemporaryUploadedFile;
+use Livewire\WithFileUploads;
 
 class UploadedFileEdit extends Component
 {
-    public UploadedFile $uploadedFile;
+    use WithFileUploads;
+
+    public UploadedFileForm $uploadedFileForm;
+
+    #[Validate(['newFile' => 'file|max:10240'])]
+    public ?TemporaryUploadedFile $newFile = null;
 
     public function mount(UploadedFile $uploadedFile): void
     {
-        $this->uploadedFile = $uploadedFile;
+        $this->uploadedFileForm->setModel($uploadedFile);
     }
 
+    /**
+     * @throws ValidationException
+     */
+    public function saveUploadedFile(): void
+    {
+        $this->uploadedFileForm->update();
+        NotificationMessage::addNotificationMessage(
+            new Item(__('File information updated.'), ItemType::SUCCESS));
+    }
+
+    public function deleteUploadedFile()
+    {
+        $fileRemoved = $this->uploadedFileForm->uploadedFile->removeFile();
+        $fileDeleted = $this->uploadedFileForm->uploadedFile->delete();
+
+        $message = 'Uploaded file was ' . ($fileRemoved ? 'removed' : 'not removed') .
+            ' and ' . ($fileDeleted ? 'deleted' : 'not deleted');
+        Log::info($message, [auth()->user(), $this->uploadedFileForm->uploadedFile]);
+        NotificationMessage::addInfoNotificationMessage(__($message));
+
+        return $this->redirect(route("uploaded-file.list"));
+    }
+
+    public function changeFile(): void
+    {
+        $this->validate();
+
+        $user = auth()->user();
+        $path = $this->newFile->storeAs('files', Str::random(5) . "_" . $this->newFile->getClientOriginalName());
+
+        $this->uploadedFileForm->uploadedFile->path = $path;
+
+        $this->uploadedFileForm->uploadedFile->name = $this->newFile->getClientOriginalName();
+        $this->uploadedFileForm->uploadedFile->mimeType = $this->newFile->getMimeType();
+
+        $this->uploadedFileForm->uploadedFile->save();
+
+        $this->uploadedFileForm->setModel($this->uploadedFileForm->uploadedFile);
+
+        Log::info('Uploaded file updated', [$user, $this->uploadedFileForm->uploadedFile]);
+        NotificationMessage::addNotificationMessage(
+            new Item(__('File uploaded.'), ItemType::SUCCESS));
+    }
 
     public function render(): \Illuminate\Contracts\View\View|\Illuminate\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\Foundation\Application
     {