+++ resources/views/components/uploaded-files-item.blade.php
@@ -15,33 +15,37 @@
         };
     }
 @endphp
-<li class="ml-5">
+<li class="ml-5 {{$file->trashed() ? 'opacity-50 line-through' : ''}}">
     <span class="@if($storerPossibleSoftDelete && $storer->trashed()) line-through text-red-700 @endif
                     @if(!$storer) font-bold @endif">
         @if($typeLink)
             <a class="underline" href="{{$typeLink}}">{{$type}}</a>
         @else
-        {{$type}}
+            {{$type}}
         @endif
     </span> -
-    @if($file->forcePublic || $file->isPublicStored()) <i class="fa-solid fa-globe"></i> @endif
+    @if($file->isPublic || $file->isPublicStored())
+        <i class="fa-solid fa-globe"></i>
+    @endif
 
     @if(!\Illuminate\Support\Facades\Storage::exists($file->path))
         <span class="text-red-700 @if($file->trashed()) line-through @endif "
-            title="{{__("File does not exit on the server.")}}">{{$file->name}}</span>
+              title="{{__("File does not exit on the server.")}}">{{$file->name}}</span>
     @else
         <span class="text-gray-500 @if($file->trashed()) line-through @endif ">
             <a class="underline" href="{{$file->getUrl()}}">{{$file->name}}</a>
         </span>
     @endif
 
-    <a href="{{route('uploaded-file.edit', $file->id)}}"
-       class="btn btn-edit max-sm:text-lg gap-2"
-       title="update uploaded file">
-        <i class="fa-solid fa-edit"></i>
-    </a>
+    @if(!$file->trashed())
+        <a href="{{route('uploaded-file.edit', $file->id)}}"
+           class="btn btn-edit max-sm:text-lg gap-2"
+           title="update uploaded file">
+            <i class="fa-solid fa-edit"></i>
+        </a>
+    @endif
 
     @if($file->uploader)
-    <span class="text-gray-500 text-xs">({{$file->uploader->name}})</span>
+        <span class="text-gray-500 text-xs">({{$file->uploader->name}})</span>
     @endif
 </li>